requirements.txt
asgiref==3.5.1
crispy-bootstrap5==0.6
Django==4.0.4
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
djangorestframework==3.13.1
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2